chore(deepsource): address Deepsource suggestions (#3436)
